backend/requirements.txt
django==6.0.3
django-bootstrap5==26.2
django-htmx==1.27.0
psycopg2-binary==2.9.11
djangorestframework==3.15.2
django-cors-headers==4.6.0
